+++ src/components/post/Comment/CommentListItem/index.tsx
@@ -29,35 +29,33 @@ const CommentListItem = ({
     onAction(action, item);
   };
 
-  return (
-    <div className={`flex flex-col gap-2 py-3 ${isLast ? "" : "border-b"}`}>
-      {isEditing ? (
-        <CommentEditForm
-          item={item}
-          onCancel={onCancelEdit}
-          onSubmit={onSubmitEdit}
-        />
-      ) : (
-        <div>
-          <div className="flex justify-between">
-            <div className="flex items-center gap-2">
-              <Image
-                src={getProfileImageUrl(item.profileImage)}
-                alt={item.nickname}
-                width={32}
-                height={32}
-                className="w-7 h-7 rounded-full"
-              />
-              <p className="text-md text-gray900">{item.nickname}</p>
-              {/* TODO: livealonedate */}
-            </div>
-            <CommentAction isMine={item.isMine} onAction={handleAction} />
-          </div>
-          <p className="text-md text-gray900 px-2">
-            {renderWithMentions(item.content)}
-          </p>
+  return isEditing ? (
+    <CommentEditForm
+      item={item}
+      onCancel={onCancelEdit}
+      onSubmit={onSubmitEdit}
+    />
+  ) : (
+    <div
+      className={`flex flex-col gap-2 px-1 py-3 ${isLast ? "" : "border-b"}`}
+    >
+      <div className="flex justify-between">
+        <div className="flex items-center gap-2">
+          <Image
+            src={getProfileImageUrl(item.profileImage)}
+            alt={item.nickname}
+            width={32}
+            height={32}
+            className="w-7 h-7 rounded-full"
+          />
+          <p className="text-md text-gray900">{item.nickname}</p>
+          {/* TODO: livealonedate */}
         </div>
-      )}
+        <CommentAction isMine={item.isMine} onAction={handleAction} />
+      </div>
+      <p className="text-md text-gray900 px-2">
+        {renderWithMentions(item.content)}
+      </p>
     </div>
   );
 };
